+++ crates/core/executor/src/minimal/unconstrained.rs
@@ -2,13 +2,13 @@
 
 use sp1_jit::JitContext;
 
-pub unsafe fn enter_unconstrained(ctx: &mut JitContext, _: u32, _: u32) -> Option<u32> {
+pub unsafe fn enter_unconstrained(ctx: &mut JitContext, _: u64, _: u64) -> Option<u64> {
     ctx.enter_unconstrained().expect("Failed to enter unconstrained mode");
 
     Some(1)
 }
 
-pub unsafe fn exit_unconstrained(ctx: &mut JitContext, _: u32, _: u32) -> Option<u32> {
+pub unsafe fn exit_unconstrained(ctx: &mut JitContext, _: u64, _: u64) -> Option<u64> {
     ctx.exit_unconstrained();
 
     Some(0)
